requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.14.0
amqp==5.3.1
APIFlask==2.3.2
apispec==6.8.1
async-timeout==5.0.1
attrs==24.3.0
Automat==24.8.1
beautifulsoup4==4.12.3
billiard==4.2.1
blinker==1.9.0
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==44.0.0
cssselect==1.2.0
defusedxml==0.7.1
filelock==3.16.1
Flask==3.1.0
Flask-HTTPAuth==4.8.0
flask-marshmallow==1.3.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.5.0
greenlet==3.1.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.10.0
itemloaders==1.3.2
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
kombu==5.4.2
lxml==5.3.0
Mako==1.3.8
MarkupSafe==3.0.2
marshmallow==3.25.1
multidict==6.1.0
packaging==24.2
parsel==1.9.1
playwright==1.49.1
prompt_toolkit==3.0.48
propcache==0.2.1
Protego==0.3.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyDispatcher==2.0.7
pyee==12.0.0
pyOpenSSL==24.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
queuelib==1.7.0
redis==5.2.1
requests==2.32.3
requests-file==2.1.0
Scrapy==2.12.0
service-identity==24.2.0
six==1.17.0
soupsieve==2.6
SQLAlchemy==2.0.37
tldextract==5.1.3
tomli==2.2.1
Twisted==24.11.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
vine==5.1.0
w3lib==2.2.1
wcwidth==0.2.13
webargs==8.6.0
Werkzeug==3.1.3
yarl==1.18.3
zope.interface==7.2
